+++ lib/honeybadger/util/http.rb
@@ -37,6 +37,12 @@ module Honeybadger
         @config = config
       end
 
+      def get(endpoint)
+        response = http_connection.get(endpoint)
+        debug { sprintf("http method=GET path=%s code=%d", endpoint.dump, response.code) }
+        response
+      end
+
       def post(endpoint, payload, headers = nil)
         response = http_connection.post(endpoint, compress(payload.to_json), http_headers(headers))
         debug { sprintf("http method=POST path=%s code=%d", endpoint.dump, response.code) }